named_entity_linking/requirements.txt
faiss_cpu==1.10.0
numpy==2.2.3
pandas==2.2.3
pytorch_metric_learning==2.8.1
scikit_learn==1.6.1
SPARQLWrapper==2.0.0
SPARQLWrapper==2.0.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0